trans-stv/requirements.txt
chattts==0.1.1
edge_tts==6.1.12
Flask==3.0.3
Flask_SocketIO==5.3.7
funasr==1.1.6
gTTS==2.5.3
numpy==1.26.4
openai==1.45.0
PyAudio==0.2.14
pydub==0.25.1
PyYAML==6.0.2
silero_vad==5.1
torch==2.4.1
torchaudio==2.4.1
Flask-SocketIO~=5.3.7
schedule==1.2.2
requests==2.32.2
charset-normalizer>=3.4.0
chardet>=5.2.0
beautifulsoup4==4.12.2
pygame==2.6.1
sounddevice==0.5.0
soundfile==0.12.1
playsound==1.3.0
langchain==0.3.2
langchain_chroma==0.1.4
langchain-community==0.3.1
langchain-core==0.3.8
langchain-openai==0.2.1
langchain-text-splitters==0.3.0
kokoro>=0.7.4
opencv-python==4.11.0.86
sentence_transformers==4.1.0
PyObjC==10.3
fastapi~=0.115.11
uvicorn[standard]~=0.32.0
websockets~=14.1
aiortc~=1.9.0
pydantic~=2.9.2